Vss2Git

+++ Source/OldCode/Archive/2007 Phasor Protocols/PhasorProtocols/IFrequencyDefinition.cs
@@ -30,7 +30,7 @@ namespace PCS.PhasorProtocols
         /// <summary>
         /// Gets or sets the df/dt offset of this <see cref="IFrequencyDefinition"/>.
         /// </summary>
-        float DfDtOffset { get; set; }
+        double DfDtOffset { get; set; }
 
         /// <summary>
         /// Gets or sets the df/dt scaling factor of this <see cref="IFrequencyDefinition"/>.
